backend/requirements.txt
torch==2.6.0
torchvision==0.21.0
flask==3.0.3
flask-cors==4.0.1
flask-jwt-extended==4.6.0
Flask-Mail==0.9.1
reportlab==4.2.5
passlib==1.7.4
opencv-python==4.10.0.84
albumentations==1.4.8
pillow==10.4.0
numpy==2.1.3
scikit-learn==1.5.2
tqdm==4.66.5
matplotlib==3.9.2
pandas==2.2.3
